requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
mlflow==2.16.0
langchain==0.2.12
langchain-groq==0.1.9
python-dotenv==1.0.1
groq==0.13.0
